Combine all progress systems
