--- Runtime/Systems/EaseSystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d184714dd778b8940b606002014f4a56
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
