--- Runtime/Systems/ElapsedSystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 833ffa034904db9429d87d22a6627739
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
