--- Runtime/Systems/LoopSystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9dcfd8c2fa6c3694abff2efafc87e0bb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
