--- Runtime/Systems/ReverseSystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8df41a4880dbec0488076cb033c3c5b1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
